Fix button not working
The QookieFix extension was calling in the content script, a function
(__cmd ) defined in the Quantcast script.
This should not have worked in the first place, due to "content script
isolation"
(https://developer.chrome.com/extensions/content_scripts#isolated_world),
for some reason it did.  Google fixed this in Chrome 71, so I use this
workaround:
https://stackoverflow.com/questions/9515704/insert-code-into-the-page-context-using-a-content-script